+++ CHANGES.rst
@@ -1,6 +1,13 @@
 Changes
 =======
 
+tip (unreleased)
+----------------
+- Fixed error that occurs when models have a foreign key pointing to a one to one field.
+
+- Allow non-integer foreign keys
+- Allow foreign keys referencing the name of the model as a string
+
 1.3.0 (2013-05-17)
 ------------------